refactor(server): split nextcloud tree shaping out of the client
